Fixed Unit Tests
