+++ tests/FamilyHubs.RequestForSupport.UnitTests/BaseWhenUsingPage.cs
@@ -1,4 +1,6 @@
 ﻿using FamilyHubs.ReferralService.Shared.Dto;
+using FamilyHubs.ReferralService.Shared.Enums;
+using FamilyHubs.ReferralService.Shared.Models;
 using FamilyHubs.RequestForSupport.Core.ApiClients;
 using FamilyHubs.SharedKernel.Identity;
 using Microsoft.AspNetCore.Http;
@@ -53,15 +55,20 @@ public abstract class BaseWhenUsingPage
         _mockReferralClientService.Setup(x => x.GetReferralById(It.IsAny<long>())).ReturnsAsync(WhenUsingTheVcsDashboard.GetReferralDto());
         _mockReferralClientService.Setup(x => x.GetReferralStatuses()).ReturnsAsync(listStatus);
 
+        List<ReferralDto> list = new() { GetReferralDto() };
+        PaginatedList<ReferralDto> pagelist = new PaginatedList<ReferralDto>(list, 1, 1, 1);
+        _mockReferralClientService.Setup(x => x.GetRequestsForConnectionByOrganisationId(It.IsAny<string>(), It.IsAny<ReferralOrderBy>(), It.IsAny<bool?>(), It.IsAny<int>(), It.IsAny<int>())).ReturnsAsync(pagelist);
+
+
 
         var displayName = "User name";
         var identity = new GenericIdentity(displayName);
-        identity.AddClaim(new Claim(FamilyHubsClaimTypes.Role, "Professional"));
+        identity.AddClaim(new Claim(FamilyHubsClaimTypes.Role, "VcsAdmin"));
         identity.AddClaim(new Claim(FamilyHubsClaimTypes.OrganisationId, "1"));
         identity.AddClaim(new Claim(FamilyHubsClaimTypes.AccountStatus, "active"));
         identity.AddClaim(new Claim(FamilyHubsClaimTypes.FullName, "Test User"));
         identity.AddClaim(new Claim(FamilyHubsClaimTypes.LoginTime, DateTime.UtcNow.ToString()));
-        identity.AddClaim(new Claim(ClaimTypes.Email, "Joe.Professional@email.com"));
+        identity.AddClaim(new Claim(ClaimTypes.Email, "vcsAdmin2.VcsAdmin@stub.com"));
         identity.AddClaim(new Claim(FamilyHubsClaimTypes.PhoneNumber, "012345678"));
         var principle = new ClaimsPrincipal(identity);
         // use default context with user
@@ -95,5 +102,51 @@ public abstract class BaseWhenUsingPage
         cfgBuilder.AddInMemoryCollection(settings);
         IConfiguration configuration = cfgBuilder.Build();
         return configuration;
+    }
+
+    public static ReferralDto GetReferralDto()
+    {
+        return new ReferralDto
+        {
+            Id = 2,
+            ReasonForSupport = "Reason For Support",
+            EngageWithFamily = "Engage With Family",
+            RecipientDto = new RecipientDto
+            {
+                Id = 2,
+                Name = "Joe Blogs",
+                Email = "JoeBlog@email.com",
+                Telephone = "078123456",
+                TextPhone = "078123456",
+                AddressLine1 = "Address Line 1",
+                AddressLine2 = "Address Line 2",
+                TownOrCity = "Town or City",
+                County = "County",
+                PostCode = "B30 2TV"
+            },
+            ReferrerDto = new ReferrerDto
+            {
+                Id = 2,
+                EmailAddress = "Bob.Referrer@email.com",
+            },
+            Status = new ReferralStatusDto
+            {
+                Name = "New",
+                SortOrder = 0
+            },
+            ReferralServiceDto = new ReferralServiceDto
+            {
+                Id = 2,
+                Name = "Service",
+                Description = "Service Description",
+                ReferralOrganisationDto = new ReferralOrganisationDto
+                {
+                    Id = 2,
+                    Name = "Organisation",
+                    Description = "Organisation Description",
+                }
+            }
+
+        };
     }
 }
